letsencrypt: Restart Boulder components on failure
This resolves the issues with the ACME tests as mentioned in commit
a75417fb13b33df0325262c8c68e80b1afbac952.

When we change the system time while these Boulder components are
running they don't seem to cope with that time warp very well and run
into a bunch of race conditions.

Even Restart=on-failure isn't very nice, but that module is only for
testing purposes, so it's not a big concern if we have a bunch of
failures during time warp.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>