Add a new PostgreSQL module for creating databases
We need to create databases on multiple occasions and
services.postgresql.initialScript doesn't quite help a lot, because the
script provided there is only run *once* for the very first start of
PostgreSQL.

So our implementation could have only created database schemas, but we
also want to set fine-grained permissions, most notably so that the
application using the respective database can't change the schema.

Our implementation is more or less centered around PostgreSQL's GRANT
statement, but without going to anal with this like specifying
column-level permissions because the latter can still be specified using
extraSql if needed.

The way how this module is supposed to be used is like this:

  headcounter.postgresql.databases = {
    shinydb.users.shiny = {};
    shinydb.neededBy = [ "shiny.service" ];
    shinydb.schemaFile = "${pkgs.shiny}/share/shiny/shiny.sql";
  };

In this case, the user shiny won't get owner privileges nor is it
possible to alter anything in the schema.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>