pkgs/mongooseim: Don't readFile from derivation
First of all, writing $out to nix-support/main-app-dir will discard its
string context and thus we only referenced the main derivation from
another derivation by accident, mainly because ${self} was also a
dependency.

Other than that, using readFile from ${self} will initiate a build of
MongooseIM during evaluation time.

This is clearly not what we want, so we now reconstruct the mainAppDir
purely in Nix, but construct it the same way within the build script so
we can fail hard whenever the mainAppDir ceases to exist.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>