tests/mongooseim: Use headcounter.postgresql
With our database abstraction in place, we can now use it to replace the
very rudimentary permission setting for the MongooseIM test.

The permissions were more lax than the default settions in
headcounter.postgresql because we didn't revoke access to the public
schema, so in theory CREATE TABLE statements could be executed by
MongooseIM.

While it may sound a bit like cracking a nut with a sledgehammer it
indeed does serve a useful purpose, mainly that we can ensure that
MongooseIM won't use statements that alter the schema.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>