mongooseim: Add patch to set default archive mode
By default the archive mode for all archiving methods that are enabled
is to *always* archive. On Headcounter we want to leave that to the
user, so that *only* those who opt in will get archiving while by
default, no archives are created.

The only exception to this is for muc, because we had logging support
for MUC previously so we won't introduce surprises.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>