xmpp: Properly configure mod_mam
We only had a very rudimentary config so far, which didn't allow users
to set their archiving preferences.

Apart from that, the max_connections parameter of PostgreSQL is now set
to 400, so we get enough room for connections from the Hydra *and* from
MongooseIM.

Related to this, we now also listen on the loopback interface, because
MongooseIM doesn't support connecting via UNIX socket.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>