mongooseim: Reload service if config has changed
So while we have ditched code reloading for now, doing reloads instead
of restarts is still a good idea if only the configuration file has
changed.

We now have a new patch which adds a way to set the new configuration
file directly within mongooseimctl. This is necessary because the
configuration file isn't in a central location but resides in a store
path.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>