modules/conditions: Add a custom command condition
Sometimes just checking for connectivity or whether a port is bindable
on some interface isn't going to be enough, for example if we want to
check whether a specific database is available on some remote server.

This allows the full flexibility by specifying a custom command that's
retried until it returns with an exit status of 0.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>