ultron: Create and check for Hydra database
This uses the new custom.command condition introduced in the previous
commit to check whether the database is available through the tunnel.

As hydra-module.nix is only creating the database if the PostgreSQL
server is running locally, we now create the database on ultron directly
using our new headcounter.postgresql.databases abstraction.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>