mongooseim: Patch in dependency on caps for PEP
The following warning is emitted in our config:

  The PEP plugin is enabled in mod_pubsub of host XXX.
  This plugin requires mod_caps to be enabled, but it isn't.

In our config however, we have enabled mod_caps, but the order in which
the modules are started is detrimental to this, because mod_pubsub is
started *before* mod_caps.

So we're now adding a hard dependency on mod_caps whenever the pubsub
module with the PEP plugin is started, which shuts up that warning.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>