services/mongooseim: Enforce config ordering rules
The host_config directive needs to be the last term in the config,
because the configuration file is evaluated in order of these terms, so
overrides in host_config suddenly won't apply anymore if they come
before the actual top-level definitions.

I've also ordered all flag values at the beginning of the config file
now, because it looks better semantically.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>