xmpp: Add special config for torservers.net
The plan for the torservers.net virtual host was to have its own MUC
instance separate from the other virtual hosts, so having our
configuration abstraction and testing infrastructure in place, we now
can easily separate modules among the virtual hosts.

This is more flexible than the upstream MongooseIM/Ejabberd
configuration, where the only way to implement this would be either to
duplicate all the module settings or to use {add, modules} and still
don't have the same flexibility because we can't omit modules, which in
turn would mean: do a basic config and then duplicate all modules that
you want to change for *every* single virtual host.

In our abstraction however, modules are simple attribute sets which we
can intersect, filter, append and whatnot.

Our escalus tests now also include an exception for torservers.net, so
we check whether the MUC that we have configured is really as it should
be, first of all in terms of service discovery but also in service
availability (we actually simulate groupchat).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @moba