acme: Prevent race conditions for permissions
We're using POSIX ACLs to set the permissions for the certificates and
keys. The behaviour of setfacl is to match the UNIX permissions as
closely as possible to the ACL, which in turn means that the permissions
we get for those files is 0640.

This in turn causes acmetool to set the permissions back to 0600 and our
path unit sets the ACL again. Unfortunately this leaves a small time
window where the file isn't accessible by the user that has previously
been given access to.

Setting this to be 0640 by default in acmetool doesn't increase the
attack vector, because the group is also "acme" and the only user in
that group is "acme" as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>