ultron: Add ssl.ca-file to lighttpd
Providing the intermediate chain along with the certificate and private key to
only ssl.pemfile doesn't really work and lighttpd wants ssl.ca-file to
be set to the intermediate chain. So in our case, let's just set it to
the full chain just to be sure (it doesn't hurt anyway).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>