xmpp: Add fallback certfile for S2S
This will be used whenever it's not targetting a specific domain and
while it's not necessarily needed I've found a few hosts that were
unable to connect due to a missing certificate.

I didn't dig too deep into *why* they're not targetting one of the
vhosts specifically but I think it's a waste of time to dig out the
server implementations of the hosts in question if we can just provide a
fallback.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>